+++ changelog.md
@@ -1,7 +1,7 @@
 # v1.1.0 
 - Ported to Geometry Dash version `2.2081`
   - Bumped Geode loader support to `v5.0.0`
-- Added in-game profile button to socials pop-up
+- Added in-game RobTop profile button to socials pop-up
 - Internal optimizations
 
 # v1.0.2